refactor: flatten project structure to repo root & fix backend startup
- Restructure: move all files from prism-desktop/ subfolder to repo root
- Fix backend compile errors: add ToSchema derives, HealthStatus.service
  field, align PhotoStatsResponse fields, fix i64 width/height casts
- Auto-create SQLite database file and working directories (uploads,
  thumbnails) on startup instead of crashing with SQLITE_CANTOPEN

+++ backend_rust/src/routes/photos/upload.rs
@@ -239,8 +239,8 @@ pub async fn upload_blob(
             .await
             .ok();
 
-        existing.width = img_info.width as i32;
-        existing.height = img_info.height as i32;
+        existing.width = img_info.width as i64;
+        existing.height = img_info.height as i64;
         existing.aspect_ratio = img_info.aspect_ratio;
         existing.hash = Some(new_hash.clone());
